refactoring to use composition not inheritance, minor fix to make integration test passing
